backend/requirements.txt
# Web Framework
sanic==23.6.0
sanic-ext==23.6.0
sanic-cors==2.3.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Migrations
alembic==1.13.0

# Authentication
pyjwt==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Excel Support
openpyxl==3.1.2
pandas==2.1.3

# Utilities
python-dotenv==1.0.0
uvloop==0.19.0
httptools==0.6.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
sanic-testing==23.6.0
